+++ app/controllers/backend/navigation/items_controller.rb
@@ -7,24 +7,28 @@ class Backend::Navigation::ItemsController < BackendController
   before_action { breadcrumb.add t('b.navigation'), backend_navigations_path }
 
   def new
-    @model = @navigation.items.new.decorate
+    @form = Backend::NavigationItemForm.new(@navigation.items.new.decorate)
   end
 
   def create
-    @model = @navigation.items.new(allowed_params).decorate
+    @form = Backend::NavigationItemForm.new(@navigation.items.new.decorate)
 
-    if @model.save
-      redirect_to backend_navigations_path,
-        notice: translate_notice(:added, :navigation_item)
+    if @form.save params[:navigation_item]
+      redirect_to backend_navigations_path, notice: translate_notice(:added, :navigation_item)
     else
       render :new
     end
   end
 
+  def edit
+    @form = Backend::NavigationItemForm.new(@model)
+  end
+
   def update
-    if @model.update_attributes allowed_params
-      redirect_to backend_navigations_path,
-        notice: translate_notice(:changes_saved)
+    @form = Backend::NavigationItemForm.new(@model)
+
+    if @form.save params[:navigation_item]
+      redirect_to backend_navigations_path, notice: translate_notice(:changes_saved)
     else
       render :edit
     end
@@ -32,28 +36,22 @@ class Backend::Navigation::ItemsController < BackendController
 
   def destroy
     @model.destroy
-    redirect_to backend_navigations_path,
-      notice: translate_notice(:deleted, :navigation_item)
+    redirect_to backend_navigations_path, notice: translate_notice(:deleted, :navigation_item)
   end
 
   private
 
   def find_navigation
-    @navigation = Navigation.find params[:navigation_id]
+    @navigation = ::Navigation.find params[:navigation_id]
   end
 
   def find_model
     @model = NavigationItem.find(params[:id]).decorate
   end
 
-  def allowed_params
-    params[:navigation_item].permit(:page_id, :extra)
-  end
-
   def translation_form
     Backend::NavigationItemTranslationForm.new(
-      navigation_item: @model,
-      translation: @model.translation(params[:translation_locale])
+      @model, @model.translation(params[:translation_locale])
     )
   end