+++ app/forms/backend/snippet_translation_form.rb
@@ -1,8 +1,30 @@
-class Backend::SnippetTranslationForm < Reform::Form
-  include Composition
+class Backend::SnippetTranslationForm < Udongo::Form
+  attr_reader :snippet, :translation
 
-  model :snippet
+  attribute :title, String
+  attribute :content, String
 
-  property :title, on: :translation
-  property :content, on: :translation
+  delegate :id, to: :snippet
+
+  def initialize(snippet, translation)
+    @snippet = snippet
+    @translation = translation
+
+    init_attribute_values(translation)
+  end
+
+  def self.model_name
+    Snippet.model_name
+  end
+
+  def persisted?
+    true
+  end
+
+  private
+
+  def save_object
+    init_object_values(@translation)
+    @translation.save
+  end
 end