+++ changelog.md
@@ -1,6 +1,8 @@
 2.0.1 - xxxx-xx-xx
 --
 * Upgrade bootstrap v4 from alpha 2 to alpha 3.
+* The reform gem is no longer used. We now use the Udongo form objects created
+  with Virtus.
 
 
 2.0.0 - 2016-07-29